Terminamos Clases Bajo Nivel
Terminas las clases de bajo nivel, todas con sus respectivos tests unitarios y un alto "Coverage" menos en mano, que habría que refactorizar para quitarle responsabilidades. Parseo hecho, feo, pero hecho. Faltan las clases de alto nivel que llamen los parser y las que manejen el juego, solo mazo estaría implementada de estas.

+++ src/main/java/edu/fiuba/algo3/modelo/comodin/SumaMultiplicador.java
@@ -6,19 +6,19 @@ import edu.fiuba.algo3.modelo.puntaje.Puntaje;
 import edu.fiuba.algo3.modelo.juego.Juego;
 
 public class SumaMultiplicador extends Comodin {
-    public SumaMultiplicador(int cantidad, Juego juego, Aleatorio probabilidad) {
+    public SumaMultiplicador(double cantidad, Juego juego, Aleatorio probabilidad) {
         super(cantidad, probabilidad, juego);
     }
 
-    public SumaMultiplicador(int cantidad, Aleatorio probabilidad) {
+    public SumaMultiplicador(double cantidad, Aleatorio probabilidad) {
         super(cantidad, probabilidad, new SinJuego());
     }
 
-    public SumaMultiplicador(int cantidad, Juego juego) {
+    public SumaMultiplicador(double cantidad, Juego juego) {
         super(cantidad, new Aleatorio(1), juego);
     }
 
-    public SumaMultiplicador(int cantidad) {
+    public SumaMultiplicador(double cantidad) {
         super(cantidad, new Aleatorio(1), new SinJuego());
     }
